+++ src/main/java/com/mw/cc/service/currencyProviders/CurrencyProvider.java
@@ -1,8 +1,7 @@
 package com.mw.cc.service.currencyProviders;
 
-import java.math.BigDecimal;
 import java.util.Map;
 
 public interface CurrencyProvider {
-   Map<String, BigDecimal> getCurrencies();
+   Map<String, Double> getCurrencies();
 }
